refactor: Restructure FRFS Structure and Common Modules and Make Confirm call Select
