+++ Backend/app/rider-platform/rider-app/Main/src/Lib/JourneyLeg/Types/Subway.hs
@@ -41,7 +41,7 @@ data SubwayLegRequestConfirmData = SubwayLegRequestConfirmData
     personId :: Id DPerson.Person,
     merchantId :: Id DMerchant.Merchant,
     merchantOperatingCityId :: Id DMOC.MerchantOperatingCity,
-    crisSdkResponse :: Maybe ApiTypes.CrisSdkResponse,
+    crisSdkResponse :: Maybe CrisSdkResponse,
     isSingleMode :: Maybe Bool,
     mbEnableOffer :: Maybe Bool,
     categorySelectionReq :: [FRFSCategorySelectionReq]
